tests/wpt/tests/html/canvas/element/line-styles/2d.line.defaults.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.line.defaults</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.line.defaults</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

  _assertSame(ctx.lineWidth, 1, "ctx.lineWidth", "1");
  _assertSame(ctx.lineCap, 'butt', "ctx.lineCap", "'butt'");
  _assertSame(ctx.lineJoin, 'miter', "ctx.lineJoin", "'miter'");
  _assertSame(ctx.miterLimit, 10, "ctx.miterLimit", "10");

});
</script>
